# backend/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Atoma API
  description: |
    REST API for Atoma - a minimalist iOS app for habit tracking, task management, and budget control.

    **Tagline**: habits, tasks, money - in one flow

    ## Authentication
    All endpoints (except auth) require a valid JWT access token in the Authorization header:
    ```
    Authorization: Bearer <access_token>
    ```

    ## Response Format
    All successful responses follow this format:
    ```json
    { "data": { ... } }
    ```

    Error responses:
    ```json
    { "error": { "code": "ERROR_CODE", "message": "..." } }
    ```
  version: 1.0.0
  contact:
    name: Atoma Support
    url: https://api.azamatbigali.online
  license:
    name: Private
servers:
  - url: https://api.azamatbigali.online/api/v1
    description: Production server
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /tasks:
    get:
      tags: [Tasks]
      summary: List Tasks
      description: Get all tasks, optionally filtered by date
      operationId: listTasks
      security:
        - bearerAuth: []
      parameters:
        - name: date
          in: query
          description: Filter by date (YYYY-MM-DD)
          schema:
            type: string
            format: date
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: List of tasks
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTasksResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Tasks]
      summary: Create Task
      description: Create a new task
      operationId: createTask
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTaskRequest'
      responses:
        '201':
          description: Task created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '400':
          $ref: '#/components/responses/ValidationError'
  /tasks/{id}:
    get:
      tags: [Tasks]
      summary: Get Task
      description: Get a single task by ID
      operationId: getTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Task details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Tasks]
      summary: Update Task
      description: Update an existing task
      operationId: updateTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTaskRequest'
      responses:
        '200':
          description: Task updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Tasks]
      summary: Delete Task
      description: Delete a task
      operationId: deleteTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Task deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'
  /tasks/{id}/toggle:
    post:
      tags: [Tasks]
      summary: Toggle Task
      description: Toggle task completion status
      operationId: toggleTask
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Task toggled
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Task'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== TRANSACTIONS ====================
  /transactions:
    get:
      tags: [Transactions]
      summary: List Transactions
      description: Get all transactions, optionally filtered by month
      operationId: listTransactions
      security:
        - bearerAuth: []
      parameters:
        - name: year
          in: query
          description: Filter by year
          schema:
            type: integer
            example: 2024
        - name: month
          in: query
          description: Filter by month (1-12)
          schema:
            type: integer
            minimum: 1
            maximum: 12
        - $ref: '#/components/parameters/limitParam'
        - $ref: '#/components/parameters/offsetParam'
      responses:
        '200':
          description: List of transactions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedTransactionsResponse'

    post:
      tags: [Transactions]
      summary: Create Transaction
      description: Create a new income or expense transaction
      operationId: createTransaction
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTransactionRequest'
      responses:
        '201':
          description: Transaction created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/ValidationError'
...
  /transactions/{id}:
    get:
      tags: [Transactions]
      summary: Get Transaction
      description: Get a single transaction by ID
      operationId: getTransaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Transaction details
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Transaction'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [Transactions]
      summary: Update Transaction
      description: Update an existing transaction
      operationId: updateTransaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateTransactionRequest'
      responses:
        '200':
          description: Transaction updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/Transaction'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [Transactions]
      summary: Delete Transaction
      description: Delete a transaction
      operationId: deleteTransaction
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/idParam'
      responses:
        '200':
          description: Transaction deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MessageResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  # ==================== RECURRING TRANSACTIONS ====================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token
  parameters:
    idParam:
      name: id
      in: path
      required: true
      description: Resource UUID
      schema:
        type: string
        format: uuid
    limitParam:
      name: limit
      in: query
      description: Number of items per page (max 100)
      schema:
        type: integer
        default: 50
        maximum: 100
    offsetParam:
      name: offset
      in: query
      description: Number of items to skip
      schema:
        type: integer
        default: 0
  responses:
    Unauthorized:
      description: Authentication required or invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: UNAUTHORIZED
              message: Invalid or expired token
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: NOT_FOUND
              message: Resource not found
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error:
              code: VALIDATION_ERROR
              message: Title is required
...
  schemas:
...
    CreateTaskRequest:
      type: object
      required: [title]
      properties:
        title:
          type: string
          maxLength: 500
        priority:
          type: string
          enum: [low, medium, high]
          default: medium
        due_date:
          type: string
          format: date
          description: Defaults to today
    UpdateTaskRequest:
      type: object
      properties:
        title:
          type: string
        is_completed:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
        due_date:
          type: string
          format: date
    Task:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        is_completed:
          type: boolean
        priority:
          type: string
          enum: [low, medium, high]
        due_date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
    PaginatedTasksResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Task'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # ==================== TRANSACTIONS ====================
    CreateTransactionRequest:
      type: object
      required: [title, amount, type]
      properties:
        title:
          type: string
          maxLength: 200
        amount:
          type: number
          format: double
          minimum: 0.01
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
          format: date
          description: Defaults to today
    UpdateTransactionRequest:
      type: object
      properties:
        title:
          type: string
        amount:
          type: number
          format: double
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
          format: date
    Transaction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        amount:
          type: number
          format: double
        type:
          type: string
          enum: [income, expense]
        category:
          type: string
        date:
          type: string
          format: date
        created_at:
          type: string
          format: date-time
...
    PaginatedTransactionsResponse:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        meta:
          $ref: '#/components/schemas/PaginationMeta'

    # ==================== RECURRING TRANSACTIONS ====================
...
    MessageResponse:
      type: object
      properties:
        data:
          type: object
          properties:
            message:
              type: string
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
    PaginationMeta:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
        total:
          type: integer